requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
# flash-attn
hydra-core
liger-kernel
numpy<2.0.0
pandas
peft
pyarrow>=19.0.0
fastparquet
pybind11
pylatexenc
pre-commit
ray[default]
tensordict>=0.8.0,<=0.9.1,!=0.9.0
torchdata
transformers
vllm==0.10.0
wandb
packaging>=20.0
uvicorn
fastapi
typer
latex2sympy2_extended
math_verify
pydantic
regex
dotenv
slist
plotly
anthropic
openai
tenacity